Move RunControl under Tester\Control namespace
