+++ src/Behat/Behat/Tester/Gherkin/OutlineTester.php
@@ -13,9 +13,9 @@ namespace Behat\Behat\Tester\Gherkin;
 use Behat\Behat\Tester\Context\ScenarioContext;
 use Behat\Gherkin\Node\OutlineNode;
 use Behat\Testwork\Tester\Context\TestContext;
+use Behat\Testwork\Tester\Control\RunControl;
 use Behat\Testwork\Tester\Exception\WrongContextException;
 use Behat\Testwork\Tester\Result\TestResults;
-use Behat\Testwork\Tester\RunControl;
 use Behat\Testwork\Tester\Tester;
 
 /**
