+++ src/Behat/Behat/Tester/Gherkin/ScenarioIsolatingTester.php
@@ -13,8 +13,8 @@ namespace Behat\Behat\Tester\Gherkin;
 use Behat\Behat\Tester\Context\ScenarioContext;
 use Behat\Testwork\Environment\EnvironmentManager;
 use Behat\Testwork\Tester\Context\TestContext;
+use Behat\Testwork\Tester\Control\RunControl;
 use Behat\Testwork\Tester\Exception\WrongContextException;
-use Behat\Testwork\Tester\RunControl;
 use Behat\Testwork\Tester\Tester;
 
 /**
