+++ src/Behat/Behat/Tester/Gherkin/ScenarioTester.php
@@ -12,10 +12,10 @@ namespace Behat\Behat\Tester\Gherkin;
 
 use Behat\Behat\Tester\Context\ScenarioContext;
 use Behat\Testwork\Tester\Context\TestContext;
+use Behat\Testwork\Tester\Control\BasicRunControl;
+use Behat\Testwork\Tester\Control\RunControl;
 use Behat\Testwork\Tester\Exception\WrongContextException;
-use Behat\Testwork\Tester\Exercise\BasicRunControl;
 use Behat\Testwork\Tester\Result\TestResults;
-use Behat\Testwork\Tester\RunControl;
 use Behat\Testwork\Tester\Tester;
 
 /**
