+++ src/Behat/Testwork/Tester/Cli/ExerciseController.php
@@ -16,9 +16,9 @@ use Behat\Testwork\Specification\SpecificationIterator;
 use Behat\Testwork\Suite\Suite;
 use Behat\Testwork\Suite\SuiteRepository;
 use Behat\Testwork\Tester\Context\ExerciseContext;
+use Behat\Testwork\Tester\Control\BasicRunControl;
 use Behat\Testwork\Tester\Exception\WrongPathsException;
 use Behat\Testwork\Tester\Exercise;
-use Behat\Testwork\Tester\Exercise\BasicRunControl;
 use Behat\Testwork\Tester\Result\ResultInterpreter;
 use Behat\Testwork\Tester\Result\TestResult;
 use Behat\Testwork\Tester\Result\TestResults;
