+++ src/Behat/Testwork/Tester/Exercise/SuiteTester.php
@@ -10,12 +10,12 @@
 
 namespace Behat\Testwork\Tester\Exercise;
 
-use Behat\Testwork\Tester\Context\TestContext;
 use Behat\Testwork\Tester\Context\SpecificationContext;
 use Behat\Testwork\Tester\Context\SuiteContext;
+use Behat\Testwork\Tester\Context\TestContext;
+use Behat\Testwork\Tester\Control\RunControl;
 use Behat\Testwork\Tester\Exception\WrongContextException;
 use Behat\Testwork\Tester\Result\TestResults;
-use Behat\Testwork\Tester\RunControl;
 use Behat\Testwork\Tester\Tester;
 
 /**
